pyproject.toml
[project]
name = "questionbank-scrapper"
version = "0.1.0"
description = "Bangladesh Govt Jobs Question Bank Crawler"
requires-python = ">=3.12"
dependencies = [
    "crawl4ai>=0.4.0",
    "streamlit>=1.32.0",
    "anthropic>=0.25.0",
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "pdfplumber>=0.10.3",
    "python-dotenv>=1.0.0",
    "aiofiles>=23.2.1",
    "lxml>=5.1.0",
    "fake-useragent>=1.5.0",
    "nest-asyncio>=1.6.0",
    "pandas>=2.2.0",
    "Pillow>=10.2.0",
    "httpx>=0.27.0",
    "tqdm>=4.66.0",
]

[tool.uv]
dev-dependencies = []
